implemented level up

+++ Assets/Scripts/Gameplay/ScrolledObjects/Enemy/EnemyLogic.cs
@@ -40,7 +40,7 @@ namespace Gameplay.ScrolledObjects.Enemy
             }
         }
 
-        public void OnHitByPlayer(ScrolledObjectView view, Action<float> hpAction, Action<int> xpAction)
+        public void OnHitByPlayer(ScrolledObjectView view, Action<int> hpAction, Action<int> xpAction)
         {
             hpAction?.Invoke(-stats.Power);
         }
