+++ Assets/Scripts/Gameplay/ScrolledObjects/IScrolledObjectLogic.cs
@@ -10,7 +10,7 @@ namespace Gameplay.ScrolledObjects
 
         public void OnHitByWeapon(ScrolledObjectView view, float damage);
 
-        public void OnHitByPlayer(ScrolledObjectView view, Action<float> hpAction, Action<int> xpAction);
+        public void OnHitByPlayer(ScrolledObjectView view, Action<int> hpAction, Action<int> xpAction);
 
         public void OnActivate(int value);
